requirements.txt
python-telegram-bot==8.1.1
sentry-sdk==0.14.2
